v1.0
I consider this release to be feature complete: you can tweak most of the settings, and print just the math problems.

Enjoy!

Feature added: only the worksheet (the math problems) are printed, and not the configuration (or debugging messages, if activated).

This has only been tested on Firefox 46.0.1 on Xubuntu 15.10, but it now does everything I wanted it to do to be (what I consider) feature complete. Given that my only assurance is "works for me!" (groan, sorry), this should be considered beta quality.

Anomaly remains: if you select "Same values in a different order" that's what happens (top numbers are scrambled, and bottom numbers are independently scrambled), regardless of the setting "Should top number always be larger than bottom number?". In other words, Scramble doesn't respect TopBigger. In many cases, just press "Same values ..." again to see if randomness worked the way you wanted the next time.

Jun 1, 2016 7:09 PM